feat: Add endpoint to get all voters
The code changes include adding a new endpoint in the `routes.js` file to handle GET requests for retrieving all voters. This endpoint is accessible to users with the roles "official" and "superAdmin". The corresponding controller function `getVoters` is implemented in the `user.controller.js` file. This feature enhances the functionality of the application by allowing authorized users to fetch all voters from the database.